==> notebooks/2_Vucar_handle_missing_value.ipynb <==
import numpy as np 
import pandas as pd

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

df = pd.read_csv("/kaggle/input/intern/car.csv")
df.head()

df = df.drop_duplicates()
df.shape

df = df.dropna(subset=['brand', 'model'])
print(df.shape)
df.isnull().sum()

df.head()

# Handle Missing Value

import json

file_path = '/kaggle/input/vucar-prepare-store-dict/final_data_dict.json'
with open(file_path, 'r', encoding='utf-8') as f:
    car_data_json = json.load(f)

def fill_missing_values(row, json_data):
    manufacture_year = str(row["manufacture_date"])
    brand = row["brand"]
    model = row["model"]

    if manufacture_year in json_data:
        brands = json_data[manufacture_year]
        if brand in brands and model in brands[brand]:
            model_info = brands[brand][model]
            for col in ["origin", "type", "seats", "gearbox", "fuel", "color"]:
                if pd.isna(row[col]):
                    row[col] = model_info.get(col, row[col])
    return row

# Áp dụng hàm điền dữ liệu vào từng hàng trong dataframe
df = df.apply(lambda row: fill_missing_values(row, car_data_json), axis=1)

df.isnull().sum()

df_cleaned = df.dropna(subset=['price'])
df_cleaned.isnull().sum()

file_path = "handled_missing.csv"
df_cleaned.to_csv(file_path, index=False)

test = pd.read_csv(file_path)
test.head()
