==> notebooks/3_Vucar_build_model.ipynb <==
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os

# I. Load Dataset

file_path = "/kaggle/input/vucar-handle-missing-value/handled_missing.csv"
df = pd.read_csv(file_path)

df.head(), df.columns

## 1. Filter brands that make up 90% of the total cars

brand_counts = df['brand'].value_counts()

total_cars = brand_counts.sum()
cumulative_percentage = (brand_counts.cumsum() / total_cars) * 100

brands_90_percent = cumulative_percentage[cumulative_percentage <= 90].index.tolist()

num_selected_brands = len(brands_90_percent)
top_brands_summary = brand_counts[brands_90_percent]

num_selected_brands, top_brands_summary

df = df[df['brand'].isin(brands_90_percent)]

df.head()

## 2. Filter only used car

used_cars_df = df[df['condition'] == 'used']

output_path = "/kaggle/working/used_car.csv"
used_cars_df.to_csv(output_path, index=False)

print(used_cars_df.shape)
used_cars_df.head()

X = used_cars_df.drop(columns=['price', "id", 'list_id', "condition"])  
y = used_cars_df['price']

X.head()

# II. Feature Engineering

brand_counts = X['brand'].value_counts()
brand_counts

## 1. Create new feature

def extract_age_features(df):
    current_year = 2024

    df['Vehicle_Age'] = current_year - df['manufacture_date']
    
    df['Mileage_per_Year'] = df['mileage_v2'] / df['Vehicle_Age']
    df['milage_with_age'] =  df.groupby('Vehicle_Age')['mileage_v2'].transform('mean')
    
    df['Mileage_per_Year_with_age'] =  df.groupby('Vehicle_Age')['Mileage_per_Year'].transform('mean')
    
    return df

def extract_other_features(df):
    
    luxury_brands =  ['Mercedes Benz', 'BMW', 'Audi', 'Porsche', 'LandRover', 
                    'Lexus', 'Jaguar', 'Bentley', 'Maserati', 'Lamborghini', 
                    'Rolls Royce', 'Ferrari', 'Aston Martin', 'Maybach']
    df['Is_Luxury_Brand'] = df['brand'].apply(lambda x: 1 if x in luxury_brands else 0)
    return df

X = extract_age_features(X)
X = extract_other_features(X)

import sklearn
print(sklearn.__version__)

# III. Training models

import pandas as pd
import numpy as np
import gc
import joblib
from sklearn.model_selection import train_test_split, KFold
from sklearn.metrics import mean_absolute_error
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder, StandardScaler, OrdinalEncoder
from sklearn.pipeline import Pipeline
from lightgbm import LGBMRegressor
from catboost import CatBoostRegressor

## 1. Handle Numerical & Categorical features

# Identify numerical and categorical features
num_features = X.select_dtypes(include=[np.number]).columns.tolist()
cat_features = X.select_dtypes(include=["object", "category"]).columns.tolist()

# Preprocessing pipelines
num_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='mean')),
    # ('scaler', StandardScaler())
])

cat_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='most_frequent')),
    ('ordinal', OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1))
])

preprocessor = ColumnTransformer(transformers=[
    ('num', num_transformer, num_features),
    ('cat', cat_transformer, cat_features)
])

## 2. Train Test Split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.05, random_state=42)

preprocessor.fit(X_train)

preprocessor_file = "/kaggle/working/preprocessor.pkl"
joblib.dump(preprocessor, preprocessor_file)
print(f"Preprocessor saved to {preprocessor_file}")

## 3. Transform Data

X_train = preprocessor.transform(X_train)
X_test = preprocessor.transform(X_test)

X_train = pd.DataFrame(X_train, columns=num_features + cat_features)
X_test = pd.DataFrame(X_test, columns=num_features + cat_features)

X_test.to_csv("X_test.csv", index=False)
y_test.to_csv("y_test.csv", index=False)

X_train

## 4. LGBM & CatBoost 5-folds CV

import numpy as np
import pandas as pd
from sklearn.model_selection import KFold
from sklearn.metrics import mean_absolute_error
import lightgbm as lgb
from catboost import CatBoostRegressor
import gc

cv = KFold(n_splits=5, shuffle=True, random_state=42)

lgb_params = {
    "boosting_type": "gbdt",
    "objective": "regression",
    "metric": "mae",
    "max_depth": 16,
    "max_bin": 255,
    "learning_rate": 0.05,
    "n_estimators": 1000,
    "colsample_bytree": 0.8,
    "verbose": -1,
    "random_state": 42,
    "device": "gpu",  # Change to 'cpu' if GPU is not available
}

catboost_params = {
    "iterations": 1000,
    "depth": 16,
    "learning_rate": 0.05,
    "loss_function": "MAE",
    "random_seed": 42,
    "verbose": 100,
    "task_type": "GPU",  # Change to 'CPU' if GPU is not available
}

lgb_oof_pred = np.zeros(X_train.shape[0])
catboost_oof_pred = np.zeros(X_train.shape[0])

lgb_fitted_models = []
catboost_fitted_models = []

for fold, (idx_train, idx_valid) in enumerate(cv.split(X_train, y_train)):
    print(f"Fold {fold + 1}")
    
    X_tr, y_tr = X_train.iloc[idx_train], y_train.iloc[idx_train]
    X_val, y_val = X_train.iloc[idx_valid], y_train.iloc[idx_valid]

    # LightGBM model
    lgb_model = lgb.LGBMRegressor(**lgb_params)
    lgb_model.fit(
        X_tr, y_tr,
        eval_set=[(X_val, y_val)],
        eval_metric="mae",
        callbacks=[
            lgb.early_stopping(stopping_rounds=100),
            lgb.log_evaluation(period=300)
        ]
    )
    lgb_fitted_models.append(lgb_model)
    lgb_oof_pred[idx_valid] = lgb_model.predict(X_val)

    # CatBoost model
    catboost_model = CatBoostRegressor(**catboost_params)
    catboost_model.fit(
        X_tr, y_tr,
        eval_set=[(X_val, y_val)],
        early_stopping_rounds=100,
    )
    catboost_fitted_models.append(catboost_model)
    catboost_oof_pred[idx_valid] = catboost_model.predict(X_val)

    gc.collect()

lgb_mae = mean_absolute_error(y_train, lgb_oof_pred)
catboost_mae = mean_absolute_error(y_train, catboost_oof_pred)

print(f"LightGBM OOF MAE: {lgb_mae:.4f}")
print(f"CatBoost OOF MAE: {catboost_mae:.4f}")

## 6. Save pretrained models

import joblib

for i, model in enumerate(lgb_fitted_models):
    joblib.dump(model, f"lgb_model_fold_{i + 1}.joblib")
for i, model in enumerate(catboost_fitted_models):
    joblib.dump(model, f"catboost_model_fold_{i + 1}.joblib")
